Enable size_t truncation warnings in most directories.
Bug: 167187
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3057921
Commit-Queue: Zhenyao Mo <zmo@chromium.org>
Auto-Submit: Peter Kasting <pkasting@chromium.org>
Owners-Override: danakj <danakj@chromium.org>
Cr-Commit-Position: refs/heads/master@{#907277}

+++ cc/metrics/compositor_frame_reporter.h
@@ -256,7 +256,9 @@ class CC_EXPORT CompositorFrameReporter {
   void AddEventsMetrics(EventMetrics::List events_metrics);
   EventMetrics::List TakeEventsMetrics();
 
-  int stage_history_size_for_testing() const { return stage_history_.size(); }
+  size_t stage_history_size_for_testing() const {
+    return stage_history_.size();
+  }
 
   void OnFinishImplFrame(base::TimeTicks timestamp);
   void OnAbortBeginMainFrame(base::TimeTicks timestamp);
